Refactor interop into separate components
